Added an embed_code option, imports-not-required. Now, by default, all code embedded and run by the doc system, must include all the imports needed to make the code runnable.
There were many codes that needed to be fixed as a result of this.
