+++ openmdao/components/tests/test_meta_model_unstructured_comp.py
@@ -1078,6 +1078,7 @@ class MetaModelTestCase(unittest.TestCase):
         self.assertTrue('step_calc' not in trig._subjacs_info[('trig.sin_x', 'trig.x2')])
 
     def test_feature_metamodel_use_fd_if_no_surrogate_linearize(self):
+        from math import sin
         from openmdao.api import SurrogateModel, MetaModelUnStructuredComp, Problem, IndepVarComp
         class SinSurrogate(SurrogateModel):
             def train(self, x, y):
