+++ openmdao/test_suite/components/sellar_feature.py
@@ -9,10 +9,11 @@ From Sellar's analytic problem.
 """
 
 import numpy as np
-
 from openmdao.api import Group, ExplicitComponent, ExecComp, IndepVarComp, \
-                         NonlinearBlockGS, ScipyKrylov, DirectSolver
-
+                         NonlinearBlockGS, ScipyKrylov, DirectSolver, EQConstraintComp, NewtonSolver
+from openmdao.test_suite.components.sellar import SellarDis1withDerivatives, \
+                         SellarDis2withDerivatives
+from openmdao.test_suite.components.double_sellar import SubSellar
 
 class SellarDis1(ExplicitComponent):
     """
@@ -234,16 +235,56 @@ class SellarNoDerivativesCS(Group):
         self.linear_solver = ScipyKrylov()
 
 
-class DoubleSellar(Group):
-
+class SellarIDF(Group):
+    """
+    Individual Design Feasible (IDF) architecture for the Sellar problem.
+    """
     def setup(self):
+        # construct the Sellar model with `y1` and `y2` as independent variables
+        dv = IndepVarComp()
+        dv.add_output('x', 5.)
+        dv.add_output('y1', 5.)
+        dv.add_output('y2', 5.)
+        dv.add_output('z', np.array([2., 0.]))
 
-        self.add_subsystem('g1', SubSellar(units=units, scaling=scaling))
-        self.add_subsystem('g2', SubSellar(units=units, scaling=scaling))
+        self.add_subsystem('dv', dv)
+        self.add_subsystem('d1', SellarDis1withDerivatives())
+        self.add_subsystem('d2', SellarDis2withDerivatives())
 
-        self.connect('g1.y2', 'g2.x')
-        self.connect('g2.y2', 'g1.x')
+        self.add_subsystem('obj_cmp', ExecComp('obj = x**2 + z[1] + y1 + exp(-y2)',
+                           x=0., z=np.array([0., 0.])))
 
-        # Converge the outer loop with Gauss Seidel, with a looser tolerance.
-        self.nonlinear_solver = NewtonSolver()
-        self.linear_solver = DirectSolver()
+        self.add_subsystem('con_cmp1', ExecComp('con1 = 3.16 - y1'))
+        self.add_subsystem('con_cmp2', ExecComp('con2 = y2 - 24.0'))
+
+        self.connect('dv.x', ['d1.x', 'obj_cmp.x'])
+        self.connect('dv.y1', ['d2.y1', 'obj_cmp.y1', 'con_cmp1.y1'])
+        self.connect('dv.y2', ['d1.y2', 'obj_cmp.y2', 'con_cmp2.y2'])
+        self.connect('dv.z', ['d1.z', 'd2.z', 'obj_cmp.z'])
+
+        # rather than create a cycle by connecting d1.y1 to d2.y1 and d2.y2 to d1.y2
+        # we will constrain y1 and y2 to be equal for the two disciplines
+
+        equal = EQConstraintComp()
+        self.add_subsystem('equal', equal)
+
+        equal.add_eq_output('y1', add_constraint=True)
+        equal.add_eq_output('y2', add_constraint=True)
+
+        self.connect('dv.y1', 'equal.lhs:y1')
+        self.connect('d1.y1', 'equal.rhs:y1')
+
+        self.connect('dv.y2', 'equal.lhs:y2')
+        self.connect('d2.y2', 'equal.rhs:y2')
+
+        # the driver will effectively solve the cycle
+        # by satisfying the equality constraints
+
+        self.add_design_var('dv.x', lower=0., upper=5.)
+        self.add_design_var('dv.y1', lower=0., upper=5.)
+        self.add_design_var('dv.y2', lower=0., upper=5.)
+        self.add_design_var('dv.z', lower=np.array([-5., 0.]), upper=np.array([5., 5.]))
+        self.add_objective('obj_cmp.obj')
+        self.add_constraint('con_cmp1.con1', upper=0.)
+        self.add_constraint('con_cmp2.con2', upper=0.)
+
